scraper/requirements.txt
rnet
psycopg2-binary
